Separate ActiveLink component from Navbar
